Code/requirements.txt
numpy
pandas
librosa
matplotlib
seaborn
tensorflow
scikit-learn
imbalanced-learn
spotipy
python-dotenv
joblib
streamlit